feat: 검색 결과가 없는 경우 대응
