+++ src/app/home/components/BoardCardList/BoardCardList.tsx
@@ -1,10 +1,12 @@
 'use client';
 
+import { SvgIcon, Typography } from '@/components';
 import { useIntersectObserver } from '@/hooks';
 import { useBoardListQuery } from '@/api/hooks';
 import { BoardCard } from '@/app/home/components';
 import { BOARDS_PER_PAGE } from '@/app/home/constants';
 import { useRegionsFilterStore } from '@/app/home/store';
+import { listGrid, noBoard } from './BoardCardList.css';
 
 const BoardCardList = () => {
   const { cityId, provinceId } = useRegionsFilterStore((state) => ({
@@ -19,13 +21,20 @@ const BoardCardList = () => {
     onIntersect: getNextPage,
   });
 
-  return (
-    <>
+  return boardList.length > 0 ? (
+    <ul className={listGrid}>
       {boardList.map((boardListItem) => (
         <BoardCard key={boardListItem.boardId} boardListItem={boardListItem} />
       ))}
       <div ref={ref}></div>
-    </>
+    </ul>
+  ) : (
+    <div className={noBoard}>
+      <SvgIcon id="info" width={40} height={40} />
+      <Typography variant="body2" as="p" color="hint">
+        검색 결과가 없습니다.
+      </Typography>
+    </div>
   );
 };
